feat: refactor density compensation to a module.
