+++ src/mrinufft/trajectories/density/__init__.py
@@ -1,0 +1,5 @@
+from .geometry_based import voronoi, voronoi_unique
+from .nufft_based import pipe
+
+
+__all__ = ["voronoi", "voronoi_unique", "pipe"]
